refactor!: shrink top-level nclutils export surface
Re-export only the high-frequency pretty-print helpers and logger from
the package root. Move every other utility behind its submodule (e.g.
nclutils.fs, nclutils.sh, nclutils.strings, nclutils.utils,
nclutils.questions, nclutils.text_processing, nclutils.network) so the
top-level namespace stays focused on the console-output hot path.

BREAKING CHANGE: imports like `from nclutils import copy_file` no
longer work. Use `from nclutils.fs import copy_file` (and equivalents
for sh, strings, utils, questions, text_processing, network).
