+++ tests/test_questions.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import questionary
 from pytest_mock import MockerFixture
 
-from nclutils import choose_multiple_from_list, choose_one_from_list
+from nclutils.questions import choose_multiple_from_list, choose_one_from_list
 
 
 def test_choose_from_list_path(mocker: MockerFixture) -> None:
